sdasaassasd

+++ FruitNindja/Assets/Scripts/Batting.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Batting : MonoBehaviour
 {
     [SerializeField] private ScoreManager _scoreManager;
-    [SerializeField] private double decreaseNumber = 100;
+    [SerializeField] private int decreaseNumber = 100;
     [SerializeField] private Animator _stoveAnimator;
     [SerializeField] private Transform _forceTransform;
